feat: global style
