+++ tpo-client/src/components/MyPage.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+
+function MyPage() {
+    return(
+        <div>
+            <h1>My page</h1>
+        {/*   create Header */}
+        {/*   create name input */}
+        {/*   add gender components */}
+        {/*   add select components */}
+        {/*   add time components */}
+        {/*   create footer button */}
+        </div>
+    )
+}
+
+export default MyPage;
