+++ tpo-client/src/index.tsx
@@ -3,14 +3,16 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import * as serviceWorkerRegistration from './serviceWorkerRegistration';
 import reportWebVitals from './reportWebVitals';
+import GlobalStyle from "./styles/GlobalStyle";
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+    document.getElementById('root') as HTMLElement
 );
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+    <React.StrictMode>
+        <GlobalStyle/>
+        <App/>
+    </React.StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
