+++ tpo-client/src/styles/GlobalStyle.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import {createGlobalStyle} from "styled-components";
+
+const GlobalStyle = createGlobalStyle`
+  * {
+    margin: 0;
+    padding: 0;;
+  }
+`
+
+export default GlobalStyle;
